requirements.txt
PyYAML==6.0.1
openai==1.35.13
python-docx==1.1.2
PyPDF2==3.0.1
markdown2==2.5.0
xhtml2pdf==0.2.16
reportlab==4.2.2
boto3==1.34.144
pandas==2.2.2
botocore==1.34.144